pyproject.toml
[tool.poetry]
name = "fbotfactory"
version = "0.1.0"
description = ""
authors = ["ilia_volodin <vollodin61@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
asyncpg = "^0.29.0"
apscheduler = "^3.10.4"
aio-scheduler = "^0.1.0"
aiohttp = "^3.9.5"
alembic = "^1.13.1"
alembic-utils = "^0.8.2"
apscheduler-di = "^0.1.0"
black = "^24.4.0"
emoji = "^2.11.0"
environs = "^11.0.0"
fastapi = "^0.110.1"
google-api-python-client = "^2.129.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
gspread-asyncio = "^2.0.0"
gspread-formatting = "^1.1.2"
gspread = "6.0"
icecream = "^2.1.3"
loguru = "^0.7.2"
redis = "^5.0.3"
oauth2client = "^4.1.3"
psycopg2 = "^2.9.9"
pyrogram = "^2.0.106"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



